# storage/api-docs/api-docs.yaml
...
openapi: "3.0.0"
info:
  title: "Upstart API"
  version: "1.0.0"
paths:
...
  /api/v1/courses:
    get:
      tags:
        - "Course"
      summary: "Get Courses"
      operationId: "Courses"
      parameters:
        - name: "category_id"
          in: "query"
          schema:
            type: "string"
            example: 1
        - name: "page"
          in: "query"
          schema:
            type: "integer"
            example: 1

        - name: "limit"
          in: "query"
          schema:
            type: "integer"
            example: 1
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Course'
        "401":
          description: "Unauthenticated"
...
  /api/v1/moderators:
    get:
      tags:
        - "Moderator"
      summary: "Get Moderators"
      operationId: "Moderator"
      parameters:
        - name: "language_code"
          in: "query"
          schema:
            type: "string"
            example: hy
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: '#/components/schemas/Moderator'
        "401":
          description: "Unauthenticated"
      security:
        - bearerAuth: [ ]
...
  /api/v1/moderators/create:
    post:
      tags:
        - "Moderator"
      operationId: "moderator"
      description: This request is only for super Admin users. In  the request headers  need set user api_token as the bearer token.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Moderator'
          application/xml:
            schema:
              $ref: '#/components/schemas/Moderator'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Moderator'
        required: true

      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: "integer"
                  first_name:
                    type: "string"
                  last_name:
                    type: "string"
                  email:
                    type: "string"
                  role_id:
                    type: "integer"
                  company_name:
                    type: "string"
                  tax_identity_number:
                    type: "integer"
                type: "object"
        "401":
          description: "Unauthenticated"
      security:
        - bearerAuth: [ ]
...
  /api/v1/resources:
    get:
      tags:
        - "Resource"
      summary: "Get all resources"
      operationId: "Resources"
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            example: 10
        - name: page
          in: query
          schema:
            type: integer
            example: 2
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    user_id:
                      type: integer
                    title:
                      type: string
                    path:
                      type: string
        "403":
          description: "Forbidden Error"
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    Course:
      required:
        - id
        - type
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
                example: 1
              type:
                type: integer
                format: int64
                example: 1
              category_id:
                type: integer
                format: int64
                example: 1
              trainer_id:
                type: integer
                format: int64
                example: 1
              title:
                type: string
                example: title course
              sub_title:
                type: string
                example: sub title course
              description:
                type: string
                example: description course
              language:
                type: integer
                example: 1
              status:
                type: integer
                example: 1
              price:
                type: integer
                example: 1.1
              max_participants:
                type: integer
                example: 1
              level:
                type: number
                example: 1.1
              currency:
                type: string
              promo_video:
                type: string
                format: binary
              cover_image:
                type: string
                format: binary
              lessons:
                type: array
                xml:
                  wrapped: true
                items:
                  type: object
                  properties:
                    title:
                      type: string
                      example: Laravel
                    course_id:
                      type: integer
                      example: 1
                    duration:
                      type: integer
                      example: 1
                    start_time:
                      type: string
                      example: 2022-10-10 10:12:00
              trainer:
                type: object
                properties:
                  first_name:
                    type: string
                    example: Trainer
                  last_name:
                    type: string
                    example: Trainer
                  bio:
                    type: string
                    example: bio
                  avatar:
                    type: string
        current_page:
          type: integer
          example: 1
        first_page_url:
          type: integer
          example: "string"
        from:
          type: integer
          example: 1
        last_page:
          type: integer
          example: 1
        last_page_url:
          type: string
          example: "http://api.loc/api/v1/courses?page=5"
        next_page_url:
          type: string
          example: "http://api.loc/api/v1/courses?page=4"
        per_page:
          type: string
          example: 1
        prev_page_url:
          example: "http://api.loc/api/v1/courses?page=2"
        to:
          type: integer
          example: 1
        total:
          type: integer
          example: 1
...
    Moderator:
      type: object
      properties:
        first_name:
          type: "string"
        last_name:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        company_name:
          type: "string"
        role_id:
          type: integer
          default: 2
        tax_identity_number:
          type: "integer"
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT